+++ FsiX.Daemon/Program.fs
@@ -6,14 +6,29 @@ open Nerdbank.Streams
 open FsiX.Daemon
 open FsiX.Daemon.Json
 
-let mkStdioJsonRpc () =
-  let stream = FullDuplexStream.Splice(Console.OpenStandardInput(), Console.OpenStandardOutput())
+open System.Net
+open System.Net.Sockets
+
+let mkStdioJsonRpc (listener: TcpListener) =
+  listener.Start()
+  printfn $"Listening on {listener.LocalEndpoint.ToString()}"
+  let client = listener.AcceptTcpClient()
+  let stream = client.GetStream()
+  //let stream = FullDuplexStream.Splice(Console.OpenStandardInput(), Console.OpenStandardOutput())
   let handler = new HeaderDelimitedMessageHandler(stream, mkJsonFormatter ())
   new JsonRpc(handler)
 
+open FSharpPlus
+let tryParseIp args = 
+  match args |> Array.tryHead |>> String.split [":"] |>> Seq.toList with 
+  | Some [ipPart; portPart] -> Some (IPAddress.Parse ipPart, Int32.Parse portPart, Array.skip 1 args)
+  | _ -> None
 [<EntryPoint>]
 let main args =
-  let rpc = mkStdioJsonRpc ()
+  let addr, port, args = 
+    tryParseIp args |> Option.defaultValue (IPAddress.Loopback, 9000, args)
+  use listener = new TcpListener(addr, port);
+  let rpc = mkStdioJsonRpc listener
   Rpc.startAndInitRpc args rpc
   rpc.Completion.GetAwaiter().GetResult()
   Console.ReadLine() |> ignore